moves helper method to elasticsearch service
