+++ src/core/server/elasticsearch/index.ts
@@ -38,4 +38,7 @@ export type {
   DeleteDocumentResponse,
 } from './client';
 export { getRequestDebugMeta, getErrorMessage } from './client';
-export { isSupportedEsServer } from './supported_server_response_check';
+export {
+  isSupportedEsServer,
+  isNotFoundFromUnsupportedServer,
+} from './supported_server_response_check';
