+++ src/core/server/saved_objects/service/lib/internal_utils.ts
@@ -13,7 +13,6 @@ import type { SavedObject } from '../../types';
 import { decodeRequestVersion, encodeHitVersion } from '../../version';
 import { SavedObjectsErrorHelpers } from './errors';
 import { ALL_NAMESPACES_STRING, SavedObjectsUtils } from './utils';
-import { isSupportedEsServer } from '../../../elasticsearch';
 
 /**
  * Checks the raw response of a bulk operation and returns an error if necessary.
@@ -142,19 +141,3 @@ export function rawDocExistsInNamespace(
     namespaces?.includes(ALL_NAMESPACES_STRING);
   return existsInNamespace ?? false;
 }
-
-/**
- * Check to ensure that a 404 response does not come from Elasticsearch
- *
- * WARNING: This is a hack to work around for 404 responses returned from a proxy.
- * We're aiming to minimise the risk of data loss when consumers act on Not Found errors
- *
- * @param response response from elasticsearch client call
- * @returns boolean 'true' if the status code is 404 and the Elasticsearch product header is missing/unexpected value
- */
-export const isNotFoundFromUnsupportedServer = (args: {
-  statusCode: number | null;
-  headers: Record<string, string | string[]> | null;
-}): boolean => {
-  return args.statusCode === 404 && !isSupportedEsServer(args.headers);
-};
